refactor: Controller 네이밍 수정 및 rest 패키지 추가
